AEA-3830 A start on unit tests for handler functions.
